GIX-2187: Default to main account when accountIdentifier is missing (#4128)
# Motivation

People want to be able to link from external websites to the ckBTC
wallet page, which has information about converting BTC to ckBTC.
This should work whether the user is signed-in or not.
When linking from an external website, the account identifier of the
user is not known so it can't be included in the URL.
To make this work we want the wallet deep-link URL work even when know
account identifier is specified.
In that case we default to the main account for the given wallet.
For consistency we do the same SNS and NSS wallets.

# Changes

1. Default to the main account when no account identifier is specified
in the `IcrcWallet` component.
2. Default to the main account when no account identifier is specified
in the `SnsWallet` component.
3. Default to the main account when no account identifier is specified
in the `NnsWallet` component.

# Tests

Adjust unit tests for each wallet where no account identifier is
specified to default to the main account instead of navigating to the
accounts page.

Fix `"should render a spinner while loading"` in `NnsWallet.spec.ts`. It
previously faked the loading state by not specifying an account
identifier. Now we put it in loading state by pausing the API.

Remove `"new transaction action should be disabled while loading"` in
`NnsWallet.spec.ts`. It did not simulate the loading state correctly and
the test below it covers the same functionality.

# Todos

- [x] Add entry to changelog (if necessary).