GIX-2187: Allow CkBTCInfoCard without account (#4133)
# Motivation

We want to allow deep linking into the ckBTC wallet page so people can
learn about BTC -> ckBTC conversion.
This should work even when users are not signed in.
The components on the wallet page currently assume that the user is
always signed in or the wallet page wouldn't be shown. So this needs to
change.

In this PR we support not being signed in for the `CkBTCInfoCard`
component by allowing to not pass the `account` prop.

When signed in:

<img width="736" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/23e13d50-92db-4027-bcb4-f8289f6783ed">

When not signed in:

<img width="755" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/f28771ca-43e7-4b1a-9b63-0b428b22adad">

A placeholder for the address QR code will be added later.

This change is not yet user visible because it's still not possible to
see the wallet page without being signed in.

# Changes

1. Change the types of `account` and `identifier` to include `undefined.
2. Load the BTC address whenever the identifier is defined, instead of
only on mount. This is for the case where the user signs in while on the
wallet page.
3. Exclude the sentence with the block explorer link if there is no
account.
4. Exclude the QR code of the BTC address if there is no account.
5. Instead of showing the BTC address, show a message to tell the user
to sign in.
6. Exclude the "Check for incoming BTC" button if there is no account.

# Tests

1. Unit tests that various elements are not rendered.
2. Unit test that the sign in message is rendered.
3. Additional unit test to make sure that setting the account after
rendering works, for the case where the user signs in on the wallet
page.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet