GIX-2176: Fix tabindex Tokens footer buttons (#4162)
# Motivation

User couldn't access the buttons in the footer with the keyboard.

Fix, use tabindex="0" instead of incrementing them.

There is an issue that the buttons are not highlighted nicely when
focused. I raised the issue with @meodai who is now working on colors.

# Changes

* Set tabindex of the TokensTableRow to `"0"`.
* Remove the prop index from the TokensTableRow.
* Typo in the class `"icon"` of the button. It should have been
`"icon-only"`. This way they come with the proper focus (and buggy)
style.
* Set the last row's tabindex of the NNS Tokens Table to `"0"` as well
and to `"-1"` for the button to avoid going there and moving directly to
the footer.

# Tests

* This logic was not tested and now it's a fixed attribute. Therefore,
no extra tests are needed.

# Todos

- [ ] Add entry to changelog (if necessary).

Not yet necessary.