test: Use the new migration test in CI (#4157)
# Motivation
The migration to stable memory needs a more tests than our
downgrade-upgrade tests provide. In previous PR, a copy of the
downgrade-upgrade test was made and refactored to allow it to test any
migration rather than just to prod and back. This test was used locally
during stable memory development but not (yet) exercised in CI. Let's
test the simplest case first - just upgrade to self - to gain confidence
that the test works as expected.

Note: This does NOT yet replace the downgrade-upgrade test. Given that
there have been many changes to the test it would be nice to know that
it works reliably in the simplest setting possible before using it for
more complicated cases.

# Changes
- Exercise the new migration test, just for a simple upgrade-to-self
with classic stable-memory-on-heap storage.

# Tests
- See CI

# Todos

- [x] Add entry to changelog (if necessary).