Transaction list placeholder (#4185)
# Motivation

[The
design](https://www.figma.com/file/x8dk5axQUf8Bi43uUsqCHc/NNS-Production-%26-Hand-over?type=design&node-id=1319-1085&mode=design&t=BiIOfY7BQjiYMj8P-4)
for the logged-out wallet page has a new placeholder for when there are
no transactions.
We should use this also when the user is logged in but there are no
transactions.

This PR introduces a simple component to use as the placeholder when
there are no transactions and uses it just for the logged-in case.


# Changes

1. Add `NoTransactions` component.
2. Use it instead of just text when there are no transactions.

# Tests

Unit tests and page objects were updated.

# Todos

- [x] Add entry to changelog (if necessary).