GIX-2187: Make SNS wallet renderable when not signed in. (#4201)
# Motivation

We want to be able to show wallet pages when the user is not logged in.
This PR makes it possible to render the `SnsWallet.svelte` page when the
user is not logged in.
This is not a user visible change because `wallet/+page.svelte` still
doesn't render any wallet when the user is not logged in.

# Changes

1. Don't load data until the user is logged in.
2. Don't render `SnsBalancesObserver` if the user is not logged in.
3. Treat `selectedAccountStore.account` as nullable because we want to
render most of the page when there is no account.
4. Use SignInGuard to render a sign in button if the user is not logged
in.
5. Render the no-transactions placeholder instead of the
SnsWalletTransactionsList component if there is no account.
6. Hide Send/Receive buttons if there is no account.

# Tests

Unit tests are added.
Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet