Feat: Dropdown Get Tokens (#4195)
# Motivation

Stop relying in the page to get test tokens.

This will be useful in the e2e tests with the new tokens page. Even
though it's not strictly needed because the user could navigate to
"/neurons" and then get the tokens.

Moreover, I believe it's a nice improvement.

# Changes

* Remove the logic of the selected token within the GetTokens.
* Add a dropdown of universes with test balance in the GetTokensModal
and use the selected universe.
* Change getTestAccountBalance to getTestIcpAccountBalance and check
only the ICP balance inside.
* Use query calls to get the balance of the test account.
* Reuse `getIcrcTokenTestAccountBalance` for balance in sns test
account.
* Rename dev service `getTokens` to `getSnsTokens`.

# Tests

* Add parameter projectName to `getSnsTokens`.
* Method `waitForOption` in Dropdown page object.
* Substitute method to get different tokens (such as `clickGetBtc`) by
one method `clickGetTokens `.
* Change implementation of `getXxxxxTokens` to wait for the option to be
present, select the option and then enter amount and submit action.
* Add optional parameter `options` to `waitFor`. The playwright default
`waitFor` waits for an element to be visible. But to wait for an
`option` we need to wait it to be in the DOM.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary. It's a dev only feature.