Feat: Remove duplicated actions cell (#4209)
# Motivation

There was an issue with the e2e tests clicking the Receive button in the
tokens table. The reason was that the selector was getting the mobile
button is hidden in desktop.

In this PR, I remove the duplicated button while maintaining the
flexibility of adding new columns.

# Changes

* Set the grid style using inline style in the TokensTableRow.svelte.
Because inline style is more than a media query, I needed to use
`!important`.

# Tests

No new tests, this is a refactor of the same current functionality.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.