Refactor: Move home page and layout (#4213)
# Motivation

There is a functionality bug because the `/` has the back button when
rendering the Tokens.

This is because it uses the layout in `/(accounts)` which is the layout
for the `/accounts` path that must have the back button.

In this PR, I moved the home page and layout from `/(u)/(accounts)/` to
`(app)/(home)/` and the accounts `/(u)/(accounts)/accounts/` to
`(u)/(list)/accounts`.

In the next PR, I will fix the bug with the functionality.

# Changes

* Move page and layout from `/(u)/(accounts)/` to `(app)/(home)/`.
* Move page and layout from `/(u)/(accounts)/accounts/` to
`(u)/(list)/accounts`.
* Fix imports in home page.

# Tests

* Fix imports in tests. Test files don't follow the file structure of
the routes. Therefore, I didn't move anything.
* Remove test `routes/root/page.spec.ts` because it's a duplication with
fewer tests of `home-page.spec.ts`.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.