Chore: Upgrade next dependencies (#4283)
# Motivation

Upgrade next dependencies after updating Candid files for sns-js.

# Changes

## Automatic Changes

* Script `npm run upgrade:next` updated the file package.lock.json

## Manual Changes

* Add `ideal_matched_participation_function` to the Aggregator type
because it's already supported in the aggregator.
* Add `ideal_matched_participation_function` and
`decentralization_swap_termination_timestamp_seconds` to the aggregator
converters.

# Tests

* Add new optional fields to quite a few mocks in our tests. I will use
this to create another PR to put all those mocks in a separate file to
have the changes centralized next time.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary. No new functionality is added.