Fix: Flaky ckbtc e2e test (#4292)
# Motivation

The ckbtc.spec e2e test is flaky and sometimes fails.

The issue seems to be that the `waitForAbsent` on the Spinner within the
balance is not waiting until it's removed.

It's weird because I would have expected it to fail if the spinner is
still there. I wasn't able to replicate it locally, so it's hard to find
the cause and the solution.

In this PR, I change the implementation of `waitForAbsent` to use
`location.waitFor`.

I will keep monitoring the CI failures.

# Changes

* Change the implementation of `waitForAbsent` in playwright element.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.