Do not hide vote in progress proposals (#4293)
# Motivation

Because "Show only proposals you can still vote for" can be unchecked or
voting actions may take a few seconds, it is expected that the proposal
will be accessible through navigation until it is complete.

1. Start from nns proposal list page
2. Uncheck “`Show only proposals you can still vote for`" if checked
3. Start voting process on proposal with id N
4. Quickly click “Older” to navigate to N-1
5. Right after navigation click “Newer”

Expected navigation to proposal N, not N+1.

# Changes

- ignore the temporal voting status (from `vote registrations store`) in
`is proposal hidden` logic. The proposal entry remains in the
`voteRegistrationStore` only till the voting is done.
- drop `VoteRegistrationStatus.complete` since it's not used anymore
(It's a relic that we still rely on it).

# Tests

- update relevant tests

# Todos

- [x] Add entry to changelog (if necessary).

# Screenshots

Issue w/ selected checkbox

https://github.com/dfinity/nns-dapp/assets/98811342/8fc2bf51-3b71-4380-83b0-aa466743a7ee

Issue w/o selected checkbox

https://github.com/dfinity/nns-dapp/assets/98811342/f6b166a3-c6af-4d5c-9a27-55c12629cbc2

After the fix

https://github.com/dfinity/nns-dapp/assets/98811342/3c5a1183-6964-4882-875c-1524c5abf41f