Chore: Add pattern to gitignore (#4320)
# Motivation

The `*.orig` extension is used for some scripts to update candid types
but it's only a temporary file. Some PRs from the bots add all the files
to git and it might include this temporary files.

In this PR, I add the extension to gitignore.

# Changes

* Add `.orig` file extension to gitignore

# Tests

I created `test.orig` file and confirmed that it was ignored by git.

# Todos

- [x] Add entry to changelog (if necessary).