Add tooltip icon to neuron state item action (#4329)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/4326, I accidentally mixed
up points 22 and 23 of
https://www.notion.so/dfinityorg/Wording-Changes-4411020138ce4d09b6605ad1fdbb9555
and put the tooltip text for the neuron state on the item action for the
dissolve delay.
In this PR I put that same text on the item action for the neuron state,
where it should be.
In the next PR I will fix the text for the dissolve delay item action.

# Changes

1. Add a tooltip icon to `NnsNeuronStateItemAction` and
`SnsNeuronStateItemAction`.
2. Pass the `token` to `SnsNeuronStateItemAction` to render in the
tooltip text.

# Tests

Unit test added.

<img width="791" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/82217201-9391-47e3-918e-0663dc569532">

# Todos

- [ ] Add entry to changelog (if necessary).
covered