Chore: Remove duplicated button styles (#4350)
# Motivation

Remove unnecessary CSS. CSS to style a link like a button was added in
the gix-components.

In this PR, remove the duplicated styles in NNS Dapp.

# Changes

* Remove mixin button.scss
* Remove CSS to use the button mixin
* Remove CSS to style an anchor tag like a button in BuyICPModal

The classes used are the same that were implemented in gix-components.
Therefore, there is no need to change them.

# Tests

I checked that all looked the same as before.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.