Feat: Poll accounts in NnsAccounts (#4411)
# Motivation

We want to pollAccounts when visiting the NnsAccounts. I don't remember
why I removed the call when the tokens page was enabled, but it's not
correct. We want to poll accounts also there.

This helps only in edge cases when the initial loading of the accounts
fails. Probably why we haven't seen any user report about it.

# Changes

* Enable `pollAccounts` in NnsAccounts when ENABLE_MY_TOKENS is enabled
also.

# Tests

* Use the mocked API data instead of setting the store. This also tests
that `pollAccounts` works as expected in NnsAccounts.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.